网盘容量

+++ back-go/service/config.go
@@ -8,7 +8,7 @@ type Config struct {
 	WebAddr          string `toml:"WebAddr"`
 	WebIndex         string `toml:"WebIndex"`
 	FilePath         string `toml:"FilePath"`
-	FileDiskTotal    int    `toml:"FileDiskTotal"`
+	FileDiskTotal    int64  `toml:"FileDiskTotal"`
 	SaveFileMultiple bool   `toml:"SaveFileMultiple"`
 }
